requirements.txt
streamlit
langchain
deep-translator
SpeechRecognition
gTTS
streamlit-audiorec
langdetect
python-dotenv
google-generativeai
